% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customscale.R
\name{customscale}
\alias{customscale}
\title{Scale training and test datasets for anomaly detection}
\usage{
customscale(DF_train, DF_test, a, b)
}
\arguments{
\item{DF_train}{The unscaled training dataset used to scale the values in 
the test dataset.}

\item{DF_test}{The unscaled test dataset scaled to the user defined interval 
[a, b].}

\item{a}{The minimum desired scaling value. This will be the minimum value of
the scaled training dataset. The test dataset minimum may be smaller than 
this value.}

\item{b}{The maximum desired scaling value. This will be the maximum value of 
the scaled training dataset. The test dataset maximum may be larger than this 
value.}
}
\value{
ScaleDF The DF_test dataset scaled to [a,b] as determined from the 
DF_train dataset
}
\description{
Scale training and test datasets for anomaly detection
}
\section{Warning}{

Requires that the test and training datasets have the same numeric features 
in the same order
}

\examples{
\dontrun{
# Select only numeric features
irisExample <- iris[ ,1:4]

# Add column to split 80\% training and 20\% test
splitIris <- irisExample \%>\%
  dplyr::mutate(Cross = sample(c(1, 2),
                              size = nrow(irisExample),
                              replace = TRUE,
                              prob = c(0.80, 0.20)))

# Obtain training dataset
irisTrain <- splitIris[splitIris$Cross == 1, ] \%>\%                                                             
  dplyr::select(-Cross)


# Obtain test dataset
irisTest <- splitIris[splitIris$Cross == 2, ] \%>\%
  dplyr::select(-Cross)

# Scale irisTrain into [0, 1] (irisTrainZO)
irisTrainZO <- customscale(irisTrain, irisTrain, 0, 1)

# Scale irisTest into [0, 1] using 
irisTest <- customscale(irisTrain, irisTest, 0 ,1)
}
}
