% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDPS_Gadget.R
\name{IDPS_Gadget}
\alias{IDPS_Gadget}
\title{Interactive anomaly detection using autoencoder neural network gadget}
\usage{
IDPS_Gadget()
}
\value{
A shiny gadget (GUI) for autoencoder anomaly detection
}
\description{
Launch the shiny gadget
}
